Refactored the backend project to separate controller functions from routes
